requirements.txt
flask==2.3.3
python-dotenv==1.0.1
firebase-admin==6.5.0
gunicorn==21.2.0
Werkzeug==2.3.7
Pillow==10.0.1